# backend/handler/handlers.gen.go
/*
 * DMARC Analyzer API
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

// Code generated by openapi-generator from api/openapi.json. DO NOT EDIT.

package handler

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

const openAPIParamsKey = "openapi.params"

type HandleDmarcChartParams struct {
	domain string
	start  *string
	end    *string
}

func GetHandleDmarcChartParams(c *gin.Context) HandleDmarcChartParams {
	value, ok := c.Get(openAPIParamsKey)
	if !ok {
		return HandleDmarcChartParams{}
	}
	params, _ := value.(HandleDmarcChartParams)
	return params
}

func bindHandleDmarcChartParams(c *gin.Context) (HandleDmarcChartParams, bool) {
	var params HandleDmarcChartParams
	params.domain = c.Param("domain")
	if params.domain == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "domain is required"})
		return params, false
	}
	if value, ok := c.GetQuery("start"); ok {
		params.start = &value
	}
	if value, ok := c.GetQuery("end"); ok {
		params.end = &value
	}
	return params, true
}

func HandleDmarcChartFromOpenAPI(c *gin.Context) {
	params, ok := bindHandleDmarcChartParams(c)
	if !ok {
		return
	}
	c.Set(openAPIParamsKey, params)
	HandleDmarcChart(c)
}

type HandleDmarcDetailParams struct {
	domain     string
	start      *string
	end        *string
	source     *string
	sourceType *string
}

func GetHandleDmarcDetailParams(c *gin.Context) HandleDmarcDetailParams {
	value, ok := c.Get(openAPIParamsKey)
	if !ok {
		return HandleDmarcDetailParams{}
	}
	params, _ := value.(HandleDmarcDetailParams)
	return params
}

func bindHandleDmarcDetailParams(c *gin.Context) (HandleDmarcDetailParams, bool) {
	var params HandleDmarcDetailParams
	params.domain = c.Param("domain")
	if params.domain == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "domain is required"})
		return params, false
	}
	if value, ok := c.GetQuery("start"); ok {
		params.start = &value
	}
	if value, ok := c.GetQuery("end"); ok {
		params.end = &value
	}
	if value, ok := c.GetQuery("source"); ok {
		params.source = &value
	}
	if value, ok := c.GetQuery("source_type"); ok {
		params.sourceType = &value
	}
	return params, true
}

func HandleDmarcDetailFromOpenAPI(c *gin.Context) {
	params, ok := bindHandleDmarcDetailParams(c)
	if !ok {
		return
	}
	c.Set(openAPIParamsKey, params)
	HandleDmarcDetail(c)
}

func HandleDomainListFromOpenAPI(c *gin.Context) {
	HandleDomainList(c)
}

type HandleDomainSummaryParams struct {
	domain string
	start  *string
	end    *string
}

func GetHandleDomainSummaryParams(c *gin.Context) HandleDomainSummaryParams {
	value, ok := c.Get(openAPIParamsKey)
	if !ok {
		return HandleDomainSummaryParams{}
	}
	params, _ := value.(HandleDomainSummaryParams)
	return params
}

func bindHandleDomainSummaryParams(c *gin.Context) (HandleDomainSummaryParams, bool) {
	var params HandleDomainSummaryParams
	params.domain = c.Param("domain")
	if params.domain == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "domain is required"})
		return params, false
	}
	if value, ok := c.GetQuery("start"); ok {
		params.start = &value
	}
	if value, ok := c.GetQuery("end"); ok {
		params.end = &value
	}
	return params, true
}

func HandleDomainSummaryFromOpenAPI(c *gin.Context) {
	params, ok := bindHandleDomainSummaryParams(c)
	if !ok {
		return
	}
	c.Set(openAPIParamsKey, params)
	HandleDomainSummary(c)
}